refactor: remove redundant `test_issue_repro` and merge relevant tests into `test_chemical_association_views`, streamline block creation logic
